requirements.txt
typer[all]
rich
pandas
google-api-python-client
google-auth-oauthlib
langchain-openai
python-dotenv
playwright
beautifulsoup4
pandas
requests
lxml
dnspython
sqlalchemy
alembic
fastapi
uvicorn
celery
redis
streamlit
plotly
pydantic
python-dateutil